pyproject.toml
[tool.poetry]
name = "dify-logger"
version = "0.1.0"
description = ""
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9"
python-dotenv = "^1.0.1"
pandas = "^2.2.2"
requests = "^2.32.3"
pandas-stubs = "^2.2.2.240514"
types-requests = "^2.32.0.20240523"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
